pyproject.toml
[project]
name = "ascii-renderer"
version = "0.1.0"
description = "High-quality ASCII art renderer using shape-based character matching"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
fast = [
    "scipy>=1.10.0",
]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
ascii-render = "ascii_renderer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
